Implement More Adjust menu
Micro-tweaking (0.1 degree rotation) is implemented in a "More" menu
which is a submenu of the Adjust menu.  All the adjust items can't fit
on a toolbar if using a small phone (like the 4S).  Flip is rarely used
and also moved to the More menu.  Exiting the More menu brings one back
to the Adjust menu.  Reset remains on the More menu, perhaps it needs to
be on both menus?